moving types to separate file
